Fix parsing for SMR Sectors File v1.06
In smrealms/smr#787, the Sectors File format was changed by adding
the `Class` property in the `[Ships]` section. As a result, MGU was
parsing the ships incorrectly in such a way that it set the number
of Cargo Holds for all ships to 0. This caused all routes to display
as "0e 0$" (no exp and no credits).